test: migrate remaining integration tests off Python-in-strings
Complete the issue #218 migration. The remaining nine subprocess test
files now install real, self-contained modules from
tests/integration/apps/ via api_source / install_api instead of
project_api_body source strings. (The sse_cors and url_reverse bodies
relied on \n and \' escaping; the autodiscovery app's secondary api.py
moves to a real module loaded with the new app_text() helper.)

Reload tests swap reload_v1 -> reload_v2 via install_api() rather than
rewriting a source string. Django settings/urls/AppConfig scaffolding
stays inline (it is config, not handlers under test).

Add test_apps_smoke.py: it imports every apps/ module and asserts each
builds a populated BoltAPI, so the modules whose subprocess tests are
Linux-only (reload, multiprocess) or artifact-gated are still validated
on every platform.

Full integration suite green: 135 passed, 5 skipped (platform/artifact),
including all 20 compression tests with brotli/zstandard installed.

Refs #218
