+++ python/tests/integration/test_runbolt_smoke.py
@@ -5,63 +5,41 @@ import time
 
 import pytest
 
+from .apps import app_source
+
 pytestmark = [pytest.mark.server_integration, pytest.mark.platform_smoke]
 
 
 def test_runbolt_serves_http_request(make_server_project):
-    project = make_server_project(
-        project_api_body="""
-        @api.get("/hello")
-        async def hello():
-            return {"message": "plain"}
-        """
-    )
+    project = make_server_project(api_source=app_source("hello"))
 
     with project.start() as server:
         response = server.get("/hello")
 
     assert response.status_code == 200
-    assert response.json() == {"message": "plain"}
+    assert response.json() == {"message": "hello"}
 
 
 def test_runbolt_dev_serves_http_request(make_server_project):
-    project = make_server_project(
-        project_api_body="""
-        @api.get("/hello")
-        async def hello():
-            return {"message": "dev"}
-        """
-    )
+    project = make_server_project(api_source=app_source("hello"))
 
     with project.start(dev=True) as server:
         response = server.get("/hello")
 
     assert response.status_code == 200
-    assert response.json() == {"message": "dev"}
+    assert response.json() == {"message": "hello"}
 
 
 @pytest.mark.skipif(not sys.platform.startswith("linux"), reason="Reload smoke runs only on Linux.")
 def test_runbolt_dev_reloads_after_file_change(make_server_project):
-    project = make_server_project(
-        project_api_body="""
-        @api.get("/version")
-        async def version():
-            return {"version": "v1"}
-        """
-    )
+    project = make_server_project(api_source=app_source("reload_v1"))
 
     with project.start(dev=True) as server:
         initial = server.get("/version")
         assert initial.json() == {"version": "v1"}
 
         time.sleep(0.3)
-        project.write_project_api(
-            """
-            @api.get("/version")
-            async def version():
-                return {"version": "v2"}
-            """
-        )
+        project.install_api(app_source("reload_v2"))
 
         payload = server.wait_for_json("/version", lambda body: body["version"] == "v2", timeout=30)
 
@@ -70,16 +48,7 @@ def test_runbolt_dev_reloads_after_file_change(make_server_project):
 
 def test_runbolt_dev_ignores_non_python_and_non_html_changes(make_server_project):
     project = make_server_project(
-        project_api_body="""
-        import os
-
-        @api.get("/reload-state")
-        async def reload_state():
-            return {
-                "pid": os.getpid(),
-                "reload_count": int(os.environ.get("DJANGO_BOLT_DEV_RELOAD_COUNT", "0")),
-            }
-        """,
+        api_source=app_source("reload_state"),
         extra_files={"testproj/runtime.log": "initial\n"},
     )
