+++ python/tests/integration/test_streaming_compression_integration.py
@@ -19,40 +19,15 @@ from __future__ import annotations
 
 import pytest
 
+from .apps import app_source
+
 pytest.importorskip("brotli", reason="brotli package required for httpx auto-decode")
 
 pytestmark = pytest.mark.server_integration
 
 
 def _make_streaming_project(make_server_project):
-    return make_server_project(
-        project_api_body="""
-        from django_bolt.responses import EventSourceResponse, StreamingResponse
-        from django_bolt.middleware import no_compress
-
-        @api.get("/events")
-        async def stream():
-            async def gen():
-                for i in range(5):
-                    yield {"i": i}
-            return EventSourceResponse(gen())
-
-        @api.get("/events/raw")
-        @no_compress
-        async def raw_stream():
-            async def gen():
-                for i in range(3):
-                    yield {"i": i}
-            return EventSourceResponse(gen())
-
-        @api.get("/text")
-        async def text_stream():
-            async def gen():
-                for chunk in ("alpha-", "beta-", "gamma"):
-                    yield chunk
-            return StreamingResponse(gen(), media_type="text/plain")
-        """,
-    )
+    return make_server_project(api_source=app_source("streaming_compression"))
 
 
 def test_default_on_brotli_sse_end_to_end(make_server_project):
