+++ python/tests/integration/test_url_reverse_server_integration.py
@@ -10,6 +10,8 @@ from __future__ import annotations
 
 import pytest
 
+from .apps import app_source
+
 pytestmark = pytest.mark.server_integration
 
 
@@ -28,26 +30,9 @@ _TEMPLATES = [
     }
 ]
 
-_API = """
-from django.template import Context, Template
-
-from django_bolt.responses import HTML
-
-
-@api.get("/missions/{mission_id}", name="get_mission")
-async def get_mission(mission_id: int):
-    return {"id": mission_id}
-
-
-@api.get("/render")
-async def render_url_tag():
-    html = Template('<a href="{% url \\'get_mission\\' mission_id=42 %}">go</a>').render(Context())
-    return HTML(html)
-"""
-
 
 def test_url_template_tag_resolves_bolt_route(make_server_project):
-    project = make_server_project(project_api_body=_API, urls_content=_URLS, templates=_TEMPLATES)
+    project = make_server_project(api_source=app_source("url_reverse"), urls_content=_URLS, templates=_TEMPLATES)
     with project.start() as server:
         body = server.wait_for_text("/render", 'href="/missions/42"')
     assert body == '<a href="/missions/42">go</a>'
